% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_helpers.R
\name{zip_llm_bundle}
\alias{zip_llm_bundle}
\title{Zip a set of files for easy sharing}
\usage{
zip_llm_bundle(files, zipfile)
}
\arguments{
\item{files}{Character vector of file paths.}

\item{zipfile}{Path to the zip file to create.}
}
\value{
The path to the created zip file.
}
\description{
Zip a set of files for easy sharing
}
