% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_schema.R
\name{generate_fake_from_schema}
\alias{generate_fake_from_schema}
\title{Generate fake data from a DB schema data.frame}
\usage{
generate_fake_from_schema(sch_df, n = 30, seed = NULL)
}
\arguments{
\item{sch_df}{data.frame from schema_from_db()}

\item{n}{rows}

\item{seed}{optional}
}
\description{
Generate fake data from a DB schema data.frame
}
