% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fake_posixct_column.R
\name{generate_fake_posixct_column}
\alias{generate_fake_posixct_column}
\title{Generate a Fake POSIXct Column}
\usage{
generate_fake_posixct_column(col, n)
}
\arguments{
\item{col}{POSIXct vector to mimic.}

\item{n}{Number of rows.}
}
\value{
POSIXct vector of length n.
}
\description{
Generate a Fake POSIXct Column
}
