% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy.R
\name{generate_fake_with_privacy}
\alias{generate_fake_with_privacy}
\title{Generate fake data using a privacy preset}
\usage{
generate_fake_with_privacy(
  data,
  n = 30,
  level = c("low", "medium", "high"),
  seed = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{n}{rows}

\item{level}{'low' | 'medium' | 'high'}

\item{seed}{optional seed}
}
\description{
Generate fake data using a privacy preset
}
