% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Prepare Input Data: Coerce to data.frame}
\usage{
prepare_input_data(data)
}
\arguments{
\item{data}{Input object}
}
\value{
A data.frame or error
}
\description{
Takes any common tabular R object and safely converts to data.frame.
Returns NULL or error if unsupported.
}
